bot/requirements.txt
aiohttp
redis
async-timeout
attrs
certifi
chardet
charset-normalizer
configparser
cycler
dateparser
et-xmlfile
idna
kiwisolver
luno-python
matplotlib
multidict
numpy
openpyxl
pandas
pyparsing
python-binance
python-dateutil
pytz
PyYAML
regex
requests
six
typing-extensions
tzlocal
ujson
urllib3
websocket-client
websockets
yarl
ccxt==4.4.4
aiofiles
tenacity
asyncpg
